fix: daily_stats avg_duration 타입 불일치 수정 (Double → Long)

+++ backend/src/main/java/com/dashboard/backend/batch/DailyStatTasklet.java
@@ -49,7 +49,7 @@ public class DailyStatTasklet implements Tasklet {
                     project, targetDate,
                     totalViews,
                     uniqueVisitors,
-                    avgDuration != null ? avgDuration : 0.0,
+                    avgDuration != null ? avgDuration.longValue() : 0L,
                     0.0  // TODO: 세션 기반 이탈률 계산
             ));
